moved template folder outside of generator folder
