--- core_lib_generator/template_core_lib/.env
@@ -1,1 +1,0 @@
-#template
