--- core_lib_generator/template_core_lib/hydra_plugins/template_core_lib/template_core_lib_searchpath.py
@@ -1,8 +1,0 @@
-from hydra.plugins.search_path_plugin import SearchPathPlugin
-from hydra.core.config_search_path import ConfigSearchPath
-
-
-class Template(SearchPathPlugin):
-    def manipulate_search_path(self, search_path: ConfigSearchPath) -> None:
-        assert isinstance(search_path, ConfigSearchPath)
-        search_path.append("template_core_lib", "pkg://template_core_lib.config")
