--- core_lib_generator/template_core_lib/setup.cfg
@@ -1,2 +1,0 @@
-[metadata]
-description_file = README.md
