--- core_lib_generator/template_core_lib/template_core_lib/__init__.py
@@ -1,2 +1,0 @@
-# Version of the core_lib
-# template_version
