--- core_lib_generator/template_core_lib/template_core_lib/config/template_core_lib.yaml
@@ -1,2 +1,0 @@
-# @package _global_
-template
