--- core_lib_generator/template_core_lib/template_core_lib/data_layers/data/db/entities/template_soft_delete_token.py
@@ -1,11 +1,0 @@
-# template_import
-
-from core_lib.data_layers.data.db.sqlalchemy.base import Base
-from core_lib.data_layers.data.db.sqlalchemy.mixins.soft_delete_mixin import SoftDeleteMixin
-from core_lib.data_layers.data.db.sqlalchemy.mixins.soft_delete_token_mixin import SoftDeleteTokenMixin
-
-
-class Template(Base, SoftDeleteMixin, SoftDeleteTokenMixin):
-    __tablename__ = 'template'
-
-    # template_column
