--- core_lib_generator/template_core_lib/template_core_lib/data_layers/data_access/template_data_access.py
@@ -1,9 +1,0 @@
-from core_lib.data_layers.data_access.data_access import DataAccess
-# template_connections_imports
-# template_function_imports
-# template_entity_imports
-
-
-class Template(DataAccess):
-# template_init
-# template_functions
