--- core_lib_generator/template_core_lib/template_core_lib/jobs/template.py
@@ -1,9 +1,0 @@
-from core_lib.jobs.job import Job
-
-
-class Template(Job):
-    def initialized(self, data_handler):
-        pass
-
-    def run(self):
-        pass
