+++ template_core_lib/.dockerignore
