+++ template_core_lib/MANIFEST.in
@@ -1,0 +1,9 @@
+include requirements.txt
+include LICENSE
+include README.md
+
+
+global-exclude *.pyc
+global-exclude __pycache__
+recursive-include hydra_plugins/* *.yaml
+recursive-include template_core_lib/* *.yaml
