+++ template_core_lib/core_lib_config.yaml
