+++ template_core_lib/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+description_file = README.md
