+++ template_core_lib/template_core_lib/__init__.py
@@ -1,0 +1,2 @@
+# Version of the core_lib
+# template_version
