+++ template_core_lib/template_core_lib/config/template_core_lib.yaml
@@ -1,0 +1,2 @@
+# @package _global_
+template
