+++ template_core_lib/template_core_lib/data_layers/data/db/entities/template.py
@@ -1,0 +1,9 @@
+# template_import
+
+from core_lib.data_layers.data.db.sqlalchemy.base import Base
+
+
+class Template(Base):
+    __tablename__ = 'template'
+
+    # template_column
