+++ template_core_lib/template_core_lib/template_core_lib.py
