+++ template_core_lib/tests/test_template.py
@@ -1,0 +1,11 @@
+import unittest
+# template_test_imports
+# sync_create_start_core_lib
+
+
+class TestTemplate(unittest.TestCase):
+    def setUpClass(self):
+        self.template_snake_core_lib = sync_create_start_core_lib()
+
+
+# template_test_functions
